Moved file locations for profile
