+++ src/App.tsx
@@ -16,7 +16,7 @@ import { IonReactRouter } from '@ionic/react-router';
 import Login from './pages/login/login';
 import Signup from './pages/signup/Signup';
 import Dashboard from './Dashboard';
-import Profile from './pages/Profile';
+import Profile from './pages/profile/Profile';
 
 /* Core CSS required for Ionic components to work properly */
 import '@ionic/react/css/core.css';
